+++ lwbtn/src/include/lwbtn/lwbtn.h
@@ -68,6 +68,11 @@ struct lwbtn_btn;
 struct lwbtn;
 
 /**
+ * \brief           Time variable type
+ */
+typedef LWBTN_CFG_TYPE_VARTYPE lwbtn_time_t;
+
+/**
  * \brief           List of button events
  * 
  */
@@ -104,27 +109,27 @@ typedef uint8_t (*lwbtn_get_state_fn)(struct lwbtn* lwobj, struct lwbtn_btn* btn
 typedef struct lwbtn_btn {
     uint16_t flags; /*!< Private button flags management */
 #if LWBTN_CFG_GET_STATE_MODE != LWBTN_GET_STATE_MODE_CALLBACK || __DOXYGEN__
-    uint8_t curr_state;         /*!< Current button state to be processed. It is used 
+    uint8_t curr_state;             /*!< Current button state to be processed. It is used 
                                     to keep track when application manually sets the button state */
-#endif                          /* LWBTN_CFG_GET_STATE_MODE != LWBTN_GET_STATE_MODE_CALLBACK || __DOXYGEN__ */
-    uint8_t old_state;          /*!< Old button state - `1` means active, `0` means inactive */
-    uint32_t time_change;       /*!< Time in ms when button state got changed last time after valid debounce */
-    uint32_t time_state_change; /*!< Time in ms when button state got changed last time */
+#endif                              /* LWBTN_CFG_GET_STATE_MODE != LWBTN_GET_STATE_MODE_CALLBACK || __DOXYGEN__ */
+    uint8_t old_state;              /*!< Old button state - `1` means active, `0` means inactive */
+    lwbtn_time_t time_change;       /*!< Time in ms when button state got changed last time after valid debounce */
+    lwbtn_time_t time_state_change; /*!< Time in ms when button state got changed last time */
 
 #if LWBTN_CFG_USE_KEEPALIVE || __DOXYGEN__
     struct {
-        uint32_t last_time; /*!< Time in ms of last send keep alive event */
-        uint16_t cnt;       /*!< Number of keep alive events sent after successful on-press detection.
+        lwbtn_time_t last_time; /*!< Time in ms of last send keep alive event */
+        uint16_t cnt;           /*!< Number of keep alive events sent after successful on-press detection.
                                     Value is reset after on-release */
-    } keepalive;            /*!< Keep alive structure */
-#endif                      /* LWBTN_CFG_USE_KEEPALIVE || __DOXYGEN__ */
+    } keepalive;                /*!< Keep alive structure */
+#endif                          /* LWBTN_CFG_USE_KEEPALIVE || __DOXYGEN__ */
 
 #if LWBTN_CFG_USE_CLICK || __DOXYGEN__
     struct {
-        uint32_t last_time; /*!< Time in ms of last successfully detected (not sent!) click event */
-        uint8_t cnt;        /*!< Number of consecutive clicks detected, respecting maximum timeout between clicks */
-    } click;                /*!< Click event structure */
-#endif                      /* LWBTN_CFG_USE_CLICK || __DOXYGEN__ */
+        lwbtn_time_t last_time; /*!< Time in ms of last successfully detected (not sent!) click event */
+        uint8_t cnt;            /*!< Number of consecutive clicks detected, respecting maximum timeout between clicks */
+    } click;                    /*!< Click event structure */
+#endif                          /* LWBTN_CFG_USE_CLICK || __DOXYGEN__ */
 
     void* arg; /*!< User defined custom argument for callback function purpose */
 
@@ -165,8 +170,8 @@ typedef struct lwbtn {
 
 uint8_t lwbtn_init_ex(lwbtn_t* lwobj, lwbtn_btn_t* btns, uint16_t btns_cnt, lwbtn_get_state_fn get_state_fn,
                       lwbtn_evt_fn evt_fn);
-uint8_t lwbtn_process_ex(lwbtn_t* lwobj, uint32_t mstime);
-uint8_t lwbtn_process_btn_ex(lwbtn_t* lwobj, lwbtn_btn_t* btn, uint32_t mstime);
+uint8_t lwbtn_process_ex(lwbtn_t* lwobj, lwbtn_time_t mstime);
+uint8_t lwbtn_process_btn_ex(lwbtn_t* lwobj, lwbtn_btn_t* btn, lwbtn_time_t mstime);
 uint8_t lwbtn_set_btn_state(lwbtn_btn_t* btn, uint8_t state);
 uint8_t lwbtn_is_btn_active(const lwbtn_btn_t* btn);