+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Develop
 
 - Add detection of the first button state to be inactive before debouncing can even start
+- Add macro `LWBTN_CFG_USE_CLICK` with default value to keep click event enabled (backward compatibility) and an option to disable *click* events
 
 ## v1.1.0