+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add detection of the first button state to be inactive before debouncing can even start
 - Add macro `LWBTN_CFG_USE_CLICK` with default value to keep click event enabled (backward compatibility) and an option to disable *click* events
+- Rework library CMake with removed INTERFACE type
 
 ## v1.1.0