+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+## v1.2.0
+
 - Add detection of the first button state to be inactive before debouncing can even start
 - Add macro `LWBTN_CFG_USE_CLICK` with default value to keep click event enabled (backward compatibility) and an option to disable *click* events
 - Rework library CMake with removed INTERFACE type